Refactor: Add handleSimple404 util function
Before this change there was a lot of repetition in model functions where they would raise a 404 error if their database query didn't find anything.
To eliminate repetition, I wrote a handleSimple404 wrapper function with a full set of unit tests which wraps one of these model functions and returns a function with 404 handling functionailty.